backend/app/requirements.txt
fastapi
uvicorn
openai
seleniumbase
bs4
python-dotenv
requests
pydantic-settings
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
python-multipart
